fixed. 修复TR种子设置分享率限制后无法显示种子列表的 BUG

+++ lib/app/home/pages/download/transmission_base_torrent.dart
@@ -30,7 +30,7 @@ class TrTorrent {
   num recheckProgress;
   int secondsDownloading;
   int secondsSeeding;
-  int seedRatioLimit;
+  double seedRatioLimit;
   int seedRatioLimited;
   int seedRatioMode;
   int sizeWhenDone;
@@ -155,7 +155,7 @@ class TrTorrent {
       recheckProgress: json['recheckProgress'] ?? 0,
       secondsDownloading: json['secondsDownloading'] ?? 0,
       secondsSeeding: json['secondsSeeding'] ?? 0,
-      seedRatioLimit: json['seedRatioLimit'] ?? 0,
+      seedRatioLimit: json['seedRatioLimit'] ?? 0.0,
       seedRatioLimited: json['seedRatioLimited'] ?? 0,
       seedRatioMode: json['seedRatioMode'] ?? 0,
       sizeWhenDone: json['sizeWhenDone'] ?? 0,
